\name{big4}
\alias{big4}
\docType{data}
\title{
big4
}
\description{
Ranking of 4 English Premier League teams from season 1992-1993 to 2012-2013.
}
\usage{data(big4)}
\format{
  A data frame with 18 observations on the following 5 variables.
  \describe{
    \item{\code{Arsenal}}{the rank of team "Arsenal"}
	\item{\code{Chelsea}}{the rank of team "Chelsea"}
	\item{\code{Liverpool}}{the rank of team "Liverpool"}
	\item{\code{Manchester.United}}{the rank of team "Manchester United"}
    \item{\code{n}}{number of observations}
  }
}
\details{
The comparative performance of the "Big Four" since the start of the English Premier League.
}
\source{
Wikipedia. (2013) http://en.wikipedia.org/wiki/Premier_League
}
\references{
Wikipedia. (2013) http://en.wikipedia.org/wiki/Premier_League
}
\examples{
data(big4)
## maybe str(big4) ; plot(big4) ...
}
\keyword{datasets}
