\name{breasttissue}
\alias{breasttissue}
\docType{data}
\title{
big4
}
\description{
Ranking of 4 breast cancer categories of 106 observations. The original dataset was a categorical one, and it was converted to the current ranking dataset by sorting the expected probabilities (computed by fitting a stepwise logistic regression) of the 4 categories.
}
\usage{data(big4)}
\format{
  A data frame with 106 observations on the following 4 variables.
  \describe{
    \item{\code{adi}}{the rank of category "adipose"}
	\item{\code{car}}{the rank of category "carcinoma"}
	\item{\code{con}}{the rank of category "connective"}
	\item{\code{new}}{the rank of cateogry "fibro-adenoma, mastopathy or glandular"}
  }
}
\details{
The rankings were generated by first fitting a stepwise logistic regression to the original dataset, and then sorted by the expected probabilities of the 4 categories.
}
\source{
Frank, A., Asuncion, A. (2010) UCI Machine Learning Repository. http://archive.ics.uci.edu/ml
}
\references{
Jossinet, J. (1996) Variability of impedivity in normal and pathological breast tissue. Medical and Biological Engineering and Computing 34, 346-350.

da Silva, J. E., de Sa, J. P., Jossinet, J. (2000) Classification of breast tissue by electrical impedance spectroscopy. Medical and Biological Engineering and Computing 38, 26-30.}
\examples{
data(breasttissue)
## maybe str(breasttissue) ; rankagg(breasttissue) ...
}
\keyword{datasets}
