% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moPMR_individual.R
\name{moPMR_individual}
\alias{moPMR_individual}
\title{The main function for probabilistic two-sample MR model for multiple outcome traits with individual-level data}
\usage{
moPMR_individual(
  xin,
  Yin,
  Zxin,
  Zyin,
  max_iterin = 1000,
  epsin = 1e-05,
  Heritability_geneexpression_threshold = 1e-04
)
}
\arguments{
\item{xin}{standarized exposure vector (e.g. gene expression in TWAS)}

\item{Yin}{complex trait matrix, each column is the standardized specific trait.}

\item{Zxin}{standardized cis-genotype matrix in eQTL data.}

\item{Zyin}{standardized cis-genotype matrix in GWAS data.}

\item{max_iterin}{The maximum iteration, which can be determined by users.}

\item{epsin}{The convergence tolerance of the absolute value of the difference  between the nth and (n+1)th log likelihood, which can be determined by users.}

\item{Heritability_geneexpression_threshold}{The threshold for the estimate of gene expression heritability explained by cis-SNPs, which can be determined by users. The causal effect pvalue will be assigned to be NA automatically if the the estimate of gene expression heritability is under this threshold.}
}
\value{
A list of estimated parameters including the p values for both the causal effect test and pleiotropy effect test. 
\item{causal_effect}{The estimates of causal effect}
\item{causal_pvalue}{The p value for the causal effect}
\item{pleiotropy_effect}{The estimates of pleiotropy effect}
\item{pleiotropy_pvalue}{The p value for the pleiotropy effect}
\item{sigma_cisSNP}{The variance estimate for the cis-SNP effect sizes}
\item{sigma_error_1}{The variance estimate of the error in eQTL data model}
\item{sigma_error_2}{The variance estimate of the error in GWAS data model}
}
\description{
moPMR_individual applies a likelihood-based approach, accounting for the correlated instruments and horizontal pleiotropy effect
}
