% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moPMR_summary.R
\name{moPMR_summary}
\alias{moPMR_summary}
\title{The main function for probabilistic two-sample MR model for multiple outcome traits with summary-level data}
\usage{
moPMR_summary(
  Zscore_1,
  Zscore_2,
  Sigma1sin,
  Sigma2sin,
  Sigmayin,
  samplen1,
  samplen2,
  max_iterin = 1000,
  epsin = 1e-05,
  Heritability_geneexpression_threshold = 1e-04
)
}
\arguments{
\item{Zscore_1}{the Zscore vector of the cis-SNP effect size vector for one specific gene in eQTL data}

\item{Zscore_2}{the Zscore matrix of the cis-SNP effect size matrix, each column for one specific gene in GWAS data}

\item{Sigma1sin}{the LD matrix in eQTL data}

\item{Sigma2sin}{the LD matrix in GWAS data,both Sigma2sin and sigma1sin are often the same from the reference panel}

\item{Sigmayin}{the corrlation matrix among multiple traits}

\item{samplen1}{the sample size of eQTL data}

\item{samplen2}{the sample size of GWAS data}

\item{max_iterin}{The maximum iteration, which can be determined by users.}

\item{epsin}{The convergence tolerance of the absolute value of the difference  between the nth and (n+1)th log likelihood, which can be determined by users.}

\item{Heritability_geneexpression_threshold}{The threshold for the estimate of gene expression heritability explained by cis-SNPs, which can be determined by users. The causal effect pvalue will be assigned to be NA automatically if the the estimate of gene expression heritability is under this threshold.}
}
\value{
A list of estimated parameters including the p values for both the causal effect test and pleiotropy effect test
\item{causal_effect}{The estimates of causal effect}
\item{causal_pvalue}{The p value for the causal effect}
\item{pleiotropy_effect}{The estimates of pleiotropy effect}
\item{pleiotropy_pvalue}{The p value for the pleiotropy effect}
\item{sigma_cisSNP}{The variance estimate for the cis-SNP effect sizes}
\item{sigma_error_1}{The variance estimate of the error in eQTL data model}
\item{sigma_error_2}{The variance estimate of the error in GWAS data model}
}
\description{
moPMR_summary applies a likelihood-based approach, accounting for the correlated instruments and horizontal pleiotropy effect
}
