\name{Probability Models for Ranking Data}
\alias{pmr-package}
\alias{pmr}
\docType{package}
\title{Probability Models for Ranking Data}

\description{
Probability models for ranking data.
}
\details{
\tabular{ll}{
Package: \tab pmr\cr
Type: \tab Package\cr
Version: \tab 1.2.2\cr
Date: \tab 2013-07-16\cr
Depends: \tab stats4, graphics\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
This package includes various probability models for ranking data. Current, only complete rankings are supported by this package.
}
\author{
Paul H. Lee and Philip L. H. Yu

Maintainer: Paul H. Lee <honglee@graduate.hku.hk>
}
\references{
Marden, J. I. (1995) Analyzing and Modeling Rank Data. Chapman and Hall.

Lee, P. H., and Yu, P. L. H. (2013) An R package for analyzing and modeling ranking data. BMC Medical Research Methodology, 13, 65.
}
\keyword{ranking data}
%%\seealso{
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%%}
%%\examples{

%%}
