% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMR_individual.R
\name{PMR_individual}
\alias{PMR_individual}
\title{The main function for probabilistic two-sample MR model with individual-level data}
\usage{
PMR_individual(yin, zin, x1in, x2in, method = "PMR_individual_Egger",
  max_iterin = 1000, epsin = 1e-05,
  Heritability_geneexpression_threshold = 1e-04)
}
\arguments{
\item{yin}{standarized exposure vector (e.g. gene expression in TWAS)}

\item{zin}{standardized complex trait vector.}

\item{x1in}{standardized cis-genotype matrix in eQTL data.}

\item{x2in}{standardized cis-genotype matrix in GWAS data.}

\item{method}{indicate the methods with Egger pleiotropy effect assumption (method="PMR_individual_Egger"), polygenic pleiotropy effect assumption with mixed chisquare approximation for the test statistics(method="PMR_individual_VC_mixchisquare") when the sample size is small or polygenic pleiotropy effect assumption with normal approximation for the test statistics(method="PMR_individual_VC_normal") when the sample size is large.}

\item{max_iterin}{The maximum iteration, which can be determined by users.}

\item{epsin}{The convergence tolerance of the absolute value of the difference  between the nth and (n+1)th log likelihood, which can be determined by users.}

\item{Heritability_geneexpression_threshold}{The threshold for the estimate of gene expression heritability explained by cis-SNPs, which can be determined by users. The causal effect pvalue will be assigned to be NA automatically if the the estimate of gene expression heritability is under this threshold.}
}
\value{
A list of estimated parameters including the p values for both the causal effect test and pleiotropy effect test. Note that the current output under polygenic pleiotropy effect assumption is only causal effect p values. 
\item{causal_effect}{The estimates of causal effect}
\item{causal_pvalue}{The p value for the causal effect}
\item{pleiotropy_effect}{The estimates of pleiotropy effect}
\item{pleiotropy_pvalue}{The p value for the pleiotropy effect}
\item{sigma_cisSNP}{The variance estimate for the cis-SNP effect sizes}
\item{sigma_error_1}{The variance estimate of the error in eQTL data model}
\item{sigma_error_2}{The variance estimate of the error in GWAS data model}
}
\description{
PMR_individual applies a likelihood-based approach, accounting for the correlated instruments and horizontal pleiotropy effect
}
