% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMR_summary_Egger.R
\name{PMR_summary_Egger}
\alias{PMR_summary_Egger}
\title{The main function for probabilistic two-sample MR model with summary data under Egger pleiotropy effect assumption}
\usage{
PMR_summary_Egger(Zscore_1, Zscore_2, Sigma1sin, Sigma2sin, samplen1,
  samplen2, lambda=0, max_iterin = 1000, epsin = 1e-05,
  Heritability_geneexpression_threshold = 1e-04)
}
\arguments{
\item{Zscore_1}{the Zscore vector of the cis-SNP effect size vector for one specific gene in eQTL data}

\item{Zscore_2}{the Zscore vector of the cis-SNP effect size vector for one specific gene in GWAS data}

\item{Sigma1sin}{the LD matrix in eQTL data}

\item{Sigma2sin}{the LD matrix in GWAS data,both Sigma2sin and sigma1sin are often the same from the reference panel}

\item{samplen1}{the sample size of eQTL data}

\item{samplen2}{the sample size of GWAS data}
  
 \item{lambda}{ The shrinkage parameter to guarantee the sparsity and positive definiteness of the estimated GWAS LD matrix using
the method in PDSCE package, the default value is 0 to indicate almost no shrikage, other choice can be empirically 
chosen to be 0.05,0.1,0.15}  

\item{max_iterin}{The maximum iteration, which can be determined by users.}

\item{epsin}{The convergence tolerance of the absolute value of the difference  between the nth and (n+1)th log likelihood, which can be determined by users.}

\item{Heritability_geneexpression_threshold}{The threshold for the estimate of gene expression heritability explained by cis-SNPs, which can be determined by users. The causal effect pvalue will be assigned to be NA automatically if the the estimate of gene expression heritability is under this threshold.}
}
\value{
A list of estimated parameters including the p values for both the causal effect test and pleiotropy effect test
\item{causal_effect}{The estimates of causal effect}
\item{causal_pvalue}{The p value for the causal effect}
\item{pleiotropy_effect}{The estimates of pleiotropy effect}
\item{pleiotropy_pvalue}{The p value for the pleiotropy effect}
\item{sigma_cisSNP}{The variance estimate for the cis-SNP effect sizes}
\item{sigma_error_1}{The variance estimate of the error in eQTL data model}
\item{sigma_error_2}{The variance estimate of the error in GWAS data model}
}
\description{
PMR_summary_Egger applies a likelihood-based approach, accounting for the correlated instruments and horizontal pleiotropy effect
}
