\name{PMR-package}
\alias{PMR-package}
\alias{PMR}
\docType{package}
\title{\packageTitle{PMR}}
\description{PMR is an R package for efficient statistical inference of two-sample MR analysis. It can account for the correlatded instruments and the horizontal pleiotropy, and can provide the accurate estimates of both causal effect and horizontal pleiotropy effect as well as the two corresponding p values.There are two main functions in PMR package, one is PMR_individual for individual level data, the other is PMR_summary_Egger for summary data.For PMR_indvidual, two pleiotropy model assumptions have been designed, one is the Egger assumption, the other is polygenic assumption (variance component model). Note that the current pleiotropy assumption for summary data is Egger, which has been implemented by function PMR_summary_Egger.}
\details{
 \tabular{ll}{
Package: \tab PMR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2019-08-04\cr
License: \tab GPL-3\cr
}
}
\author{
   Zhongshang Yuan, Xiang Zhou
  
  Maintainer: Zhongshang Yuan <yuanzhongshang@sdu.edu.cn>
}
\references{
Yuan, Z., Zhu, H., Zeng, P., Yang, S., Sun, S., Yang, C., Liu, J., and Zhou, X. (2019). Testing and controlling for horizontal pleiotropy with the probabilistic mendelian randomization in transcriptome-wide association studies. bioRxiv pp. 691014.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}

